Removed half pixel shift of extent, and distinction/confustion of the PIXEL_OR_AREA metadata item
After a bit of digging I've come come to
the conclusion that this is currently being
correctly handled by GDAL such that
getting the origin or corner cordinates
will return the actual plotting/data extent.
As such the returned extent should be a simple
reordering of the bounds.

I verified this in ArcMap and QGis against a
couple of datasets.

For context see:
https://trac.osgeo.org/gdal/ticket/3837#comment:9
and
https://trac.osgeo.org/gdal/wiki/rfc33_gtiff_pixelispoint