Adds function that rounds window to include full expanse of intersecting tiles.
When finding a window for clipping a subset of a raster, it is sometimes useful for the window to evenly split on the borders of the raster's tiles. For example, if a subset is clipped out of a larger raster, if that subset raster will end up being multiplied against or otherwise compared to the larger raster, having the tiles line up makes for a cleaner operation. In addition, if running some kind of block statistics of a larger raster and a subset, having tiles that are equal makes it easier to compare differently shaped, but identically tiled rasters.