Fix dtype check in dataset write
It looks like the previously implemented data type check was not ever
being entered, due to checking for a 'type' attribute on a string. Since
the _init_dtype attribute is now always a string, remove this check and
related handling, so that now dtypes are checked for validity, providing
a more useful error message in case of an invalid dtype. Fixes #1008.