Populate `pan{Src|Dst}Bands` indexes when reprojecting
This is addressing issue #1350. Basically warping for cases other than

- Warping first band
- Warping all bands

didn't work, because of the way warping structure was configured. With this
change we actually populate `pan{Src|Dst}Bands` parameters properly.

Also when destination is an `ndarray` and there are multiple bands being
processed destination bands are assumed to be `1,2,..N` if the output size
matches the number of bands being processed simultaneously. So you can extract
say bands `4,7,19` into a `3xHxW` `ndarray`.