Populate `pan{Src|Dst}Bands` indexes when reprojecting (#1424)
* Populate `pan{Src|Dst}Bands` indexes when reprojecting

This is addressing issue #1350. Basically warping for cases other than

- Warping first band
- Warping all bands

didn't work, because of the way warping structure was configured. With this
change we actually populate `pan{Src|Dst}Bands` parameters properly.

Also when destination is an `ndarray` and there are multiple bands being
processed destination bands are assumed to be `1,2,..N` if the output size
matches the number of bands being processed simultaneously. So you can extract
say bands `4,7,19` into a `3xHxW` `ndarray`.

* Being explicit about resampling strategy in the test

Use `nearest` resampling strategy in `test_issue1350`, in case the difference in
test failure across GDAL versions is due to that.

* Update test for issue 1350 following feedback

* Skip reproject test for gdal 2.2.x series

There was a regression in 2.2.x series of GDAL, where warping would use
`panDstBands` instead of `panSrcBands` when warping into a 1 band destination.
Regression was fixed in 2.2.4 release of GDAL.