Use the right dtype when checking nodataval (#2100)
* Use the right dtype when checking nodataval

Use the dtype of the first dataset if dtype is None. If dtype is not
None, then dt == dtype.

* Use np.issubdtype to check dtypes.

* Add test for out of range nodata.

* Only query float dtype when nodataval is not NaN.

* Modify warning phrasing.

At this point the nodataval could either be set by the input files or by
the user. Use a more general phrasing to just convey the fact that the
nodata value is out of range of the dtype.