Allow AWSSession construction without boto3
Should be able to use AWSSession even when boto3 is not installed so
long as unsigned access is used.

AWSSession is responsible to configuration of things like region and
endpoint url, so it's needed even when accessing public buckets without
relying on environment variables. `boto3` dependency is only needed for
obtaining credentials though.