Fix calculation of bounds in transform.array_bounds (#2790)
Fixes #2787.

The method of computing raster bounds in `rasterio.transform.array_bounds`
assumes a standard GIS orientation of the image - origin bottom left, north
up, east right. If the image is rotated in any way, the returned bounds
may be in the correct order (e.g. `east` is west of `west`), or may
be entirely incorrect.

Replace the current implementation with the more general calculation used
in DatasetBase.bounds. Add an array of tests to both implementations,
passing in various transforms.