Fix geojson_type click option implementation (#3391)
According to several issues upstream (e.g., [1], [2]), creating multiple
options that save to the same destination is not documented or
supported. Somewhere in 8.2.x, this also broke, as it depended on
evaluation order in some way.

Instead, save the options to different name, and use a callback function
on all of them to verify that the result is exclusive, and follows the
default.

[1] https://github.com/pallets/click/issues/1688#issuecomment-707918832
[2] https://github.com/pallets/click/issues/2321#issuecomment-1184398784